Fix Ajax for sending data from Javascript to ActiveRecord

+++ db/migrate/20150602121017_create_games.rb
@@ -1,7 +1,7 @@
 class CreateGames < ActiveRecord::Migration
   def change
     create_table :games do |game|
-      game.integer :time_taken
+      game.float :time_taken
       game.boolean :done
       game.timestamps
     end
